Errors universe restructured
